# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UniHub Workshop API
  version: 1.0.0
  description: "REST API specification for the **UniHub Workshop** system — university\
    \ workshop\nregistration & check-in platform. This document is the authoritative\
    \ contract\nbetween backend (NestJS Modular Monolith) and frontend (Next.js Web\
    \ Portal +\nExpo Mobile App).\n\n## Conventions\n\n- **Versioning:** URL-based,\
    \ prefix `/api/v1`.\n- **Authentication:** JWT Bearer (15-minute access token)\
    \ + refresh token (7 days).\n  Refresh token transport is **client-type specific**\
    \ (see `authentication.md` INV-02):\n  - **Web client:** `refresh_token` is set\
    \ as an `HttpOnly; SameSite=Strict` cookie\n    on `/auth/login` and `/auth/refresh`\
    \ responses. JavaScript cannot read it.\n    Access token is returned in the JSON\
    \ body and kept **in memory only** (not localStorage).\n  - **Mobile client:**\
    \ `refresh_token` is also included in the JSON response body\n    (in addition\
    \ to the cookie) because mobile has no cookie jar — app stores it in\n    Android\
    \ Keystore / iOS Keychain.\n- **Idempotency:** `POST /registrations` and `POST\
    \ /payments` require an\n  `X-Idempotency-Key: <UUID v4>` **HTTP header** (client-generated,\
    \ reuse on retry).\n  Server uses 3-state lifecycle (`in_progress` / `completed`\
    \ / `unresolved`).\n  See ADR-08, ADR-15.\n- **Optimistic Locking:** Admin updates\
...
    \ `VND`).\n"
  contact:
...
  license:
...
servers:
...
security:
- bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags:
      - Auth
      summary: Login (student or staff)
      description: "Returns a JWT access token (15 min) and a refresh token (7 days).\n\
        Server distinguishes account type via the `accountType` field.\n\n## Refresh\
        \ token transport (`authentication.md` INV-02)\n\nBoth client types receive\
        \ the refresh token, but via different mechanisms:\n\n**Web client:**\n- `access_token`\
        \ in JSON body — kept in JS memory, **not** `localStorage`.\n- `refresh_token`\
        \ set as `Set-Cookie: refresh_token=...; HttpOnly; SameSite=Strict; Path=/auth/refresh;\
        \ Max-Age=604800`.\n  JavaScript cannot read it. XSS-safe.\n\n**Mobile client\
        \ (checkin_staff):**\n- Same `access_token` in JSON body.\n- `refresh_token`\
        \ **also** included in the JSON body (mobile has no cookie jar).\n  App stores\
        \ it in Android Keystore / iOS Keychain.\n\nFor role `checkin_staff`, the\
        \ response `user.allowedWorkshopIds` populates\nthe mobile app's `appSession.allowedWorkshopIds`.\n"
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: '`HttpOnly; SameSite=Strict; Path=/auth/refresh` cookie
                carrying the

                refresh token. **Web clients only** — mobile receives `refresh_token`

                in the response body. JavaScript cannot access this header value.

                '
              schema:
                type: string
                example: refresh_token=eyJ...; HttpOnly; SameSite=Strict; Path=/auth/refresh;
                  Max-Age=604800
            X-RateLimit-Limit:
              $ref: '#/components/headers/RateLimit-Limit'
            X-RateLimit-Remaining:
              $ref: '#/components/headers/RateLimit-Remaining'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimit-Reset'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                student:
                  summary: Student login (web)
                  value:
                    accessToken: eyJhbGc...
                    tokenType: Bearer
                    expiresIn: 900
                    role: student
                    user:
                      id: '21127001'
                      role: student
                      fullName: Nguyễn Văn A
                      email: 21127001@uni.edu.vn
                    refreshToken: null
                checkinStaffMobile:
                  summary: Check-in staff login (mobile — refresh_token in body)
                  value:
                    accessToken: eyJhbGc...
                    tokenType: Bearer
                    expiresIn: 900
                    role: checkin_staff
                    user:
                      id: 550e8400-e29b-41d4-a716-446655440000
                      role: checkin_staff
                      fullName: Trần Thị B
                      email: tran.b@uni.edu.vn
                      allowedWorkshopIds:
                      - 3f2a1c5e-...
                      - 7b9d2e1f-...
                    refreshToken: eyJhbGc...
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account disabled (`auth.account_disabled`)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh access token
      description: 'Exchange a valid refresh token for a new access + refresh token
        pair.

        Refresh token is rotated on each call (one-time use).


        ## Transport (`authentication.md` INV-02)


        **Web client:** Send with no extra header — the browser automatically

        sends the `refreshToken` HttpOnly cookie (set on `/auth/login`).

        The `Path=/auth/refresh` attribute ensures the cookie is only sent

        to this endpoint, not leaked to other API calls.


        **Mobile client:** Send `refreshToken` in the JSON request body

        (mobile has no cookie jar). Server accepts either source.


        On success, the old refresh token is invalidated and a new one is issued

        via the same transport (cookie for web, body for mobile).

        '
      operationId: refresh
      security: []
      requestBody:
        description: 'Required for **mobile clients only**. Web clients omit the body;

          the server reads the `refreshToken` HttpOnly cookie automatically.

          '
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  description: Mobile only — the stored refresh token from login response
                    body
      responses:
        '200':
          description: Refresh successful
          headers:
            Set-Cookie:
              description: Rotated refresh token cookie (web clients)
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 'Refresh token invalid (`auth.refresh_invalid`) or expired

            (`auth.refresh_expired`).

            '
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /auth/logout:
    post:
      tags:
      - Auth
      summary: Logout
      description: 'Revokes the current refresh token by adding it to a Redis denylist

        (TTL = remaining lifetime). Access token expires naturally in ≤15 min.

        '
      operationId: logout
      responses:
        '204':
          description: Logout successful
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags:
      - Auth
      summary: Get current user profile
      operationId: getMe
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /device-tokens:
    post:
      tags:
      - Device Tokens
      summary: Register a push notification token
      description: 'Upserts a FCM (Android) or APNs (iOS) device token for the current

        student. If the same `token` already exists for another student

        (e.g., shared device, app reinstall), the binding is moved to the

        current student (`isActive=true`, `lastSeen=now()`).


        **Role:** `student` only.

        '
      operationId: registerDeviceToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceTokenRequest'
      responses:
        '201':
          description: Token registered (new)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceToken'
        '200':
          description: Token updated (upsert)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceToken'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /device-tokens/{token}:
    delete:
      tags:
      - Device Tokens
      summary: Deactivate a push token
      description: 'Sets `isActive=false` (soft delete to keep audit trail). Called
        when

        the user logs out on a device.

        '
      operationId: deactivateDeviceToken
      parameters:
      - name: token
        in: path
        required: true
        description: The FCM/APNs token string
        schema:
          type: string
      responses:
        '204':
          description: Token deactivated
        '404':
          $ref: '#/components/responses/NotFound'
...
  /admin/checkin-staff/{userId}/assign-workshops:
    post:
      tags:
      - Check-in Staff Admin
      summary: Assign workshops to a check-in staff member
      description: |
        Replaces the full set of assigned workshops for a check-in staff
        member (not incremental). Previously assigned workshops not in
        the new list are unlinked.
        **Role:** `btc`.
      operationId: adminAssignCheckinWorkshops
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignWorkshopsRequest'
      responses:
        '200':
          description: Workshops assigned
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/checkin-staff/{userId}/workshops:
    get:
      tags:
      - Check-in Staff Admin
      summary: Get workshops assigned to a check-in staff member
      operationId: adminGetCheckinWorkshops
      parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: List of assigned workshop IDs
          content:
            application/json:
              schema:
                type: object
                properties:
                  workshopIds:
                    type: array
                    items:
                      type: string
                      format: uuid
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Short-lived access token (15 min). RS256 signed. Decoded claims:

        `sub` (studentId or staff UUID), `role`, `user_type`,

        `allowedWorkshopIds` (checkin_staff only), `iat`, `exp`.

        Verified by public key — no DB lookup at Layer ① (authentication.md INV-01).

        '
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
      description: HMAC-SHA256 signature of the request body. Per-gateway secret.
  headers:
    RateLimit-Limit:
      description: Quota of the narrowest tier hit (per-user, per-endpoint, or per-ip)
      schema:
        type: integer
    RateLimit-Remaining:
      description: Remaining requests in the current window
      schema:
        type: integer
    RateLimit-Reset:
      description: Unix timestamp when the window slides forward
      schema:
        type: integer
    Retry-After:
      description: Seconds to wait before retrying (only with 429 / 503)
      schema:
        type: integer
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: Authenticated but lacks permission for this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found, or hidden by row-level filter
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    TooManyRequests:
      description: "Rate limit exceeded (`rate-limiting.md`). The `error.tier` field\
        \ indicates\nwhich tier tripped:\n- `T1_ip` — 60 req/60s per IP (unauthenticated\
        \ endpoints only)\n- `T2_user` — 30 req/60s per user (all authenticated endpoints)\n\
        - `T3_user_workshop` — 5 req/60s per user per workshop\n  (`POST /registrations`\
        \ and `POST /payments` only)\n\nClient should back off for exactly `Retry-After`\
        \ seconds (calculated\nprecisely from the sliding window, not hard-coded)\
        \ then retry.\nT3 violation on workshop A does **not** block registration\
        \ for workshop B.\n"
      headers:
        Retry-After:
          $ref: '#/components/headers/Retry-After'
        X-RateLimit-Limit:
          description: Limit of the narrowest tier that was tripped
          schema:
            type: integer
        X-RateLimit-Remaining:
          schema:
            type: integer
            example: 0
        X-RateLimit-Reset:
          description: Unix timestamp when the oldest entry in the sliding window
            expires
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            success: false
            error:
              code: rate_limit_exceeded
              message: Bạn đã gửi quá nhiều lần đăng ký workshop này. Vui lòng thử lại sau 23 giây.
            meta:
              requestId: req-abc123
              timestamp: '2026-05-09T12:00:00+07:00'
              apiVersion: v1
            tier: T3_user_workshop
            retryAfter: 23
  schemas:
    ErrorEnvelope:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          enum:
          - false
          description: Always false for error responses
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              description: Stable error code clients can switch on
              example: registration.workshop_full
            message:
              type: string
              description: Human-readable error message safe for client display
              example: Workshop đã đầy chỗ
            fieldErrors:
              type: array
              items:
                $ref: '#/components/schemas/FieldError'
              description: Field-level validation details when applicable
        meta:
          $ref: '#/components/schemas/RequestMeta'
...
    RequestMeta:
      type: object
      description: Metadata attached to every API response for tracing
      required:
      - requestId
      - timestamp
      - apiVersion
      properties:
        requestId:
          type: string
          description: Correlation ID for request tracing
          example: req-abc123
        timestamp:
          type: string
          format: date-time
          description: Response timestamp (RFC 3339)
        apiVersion:
          type: string
          description: API version tag
          example: v1
        processingMs:
          type: integer
          description: Server-side processing duration in milliseconds
          example: 42
    FieldError:
      type: object
      description: Field-level validation error details
      required:
      - field
      - rule
      - message
      properties:
        field:
          type: string
          description: Field path that failed validation
          example: email
        rule:
          type: string
          description: Validation rule identifier
          example: isEmail
        message:
          type: string
          description: Human-readable validation message
          example: Invalid email format
        received:
          description: Raw input value (when safe to expose)
...
    LoginRequest:
      type: object
      required:
      - accountType
      - password
      description: 'Login body. Field `studentId` is required when `accountType=student`;

        field `email` is required when `accountType=staff`. This split avoids a

        generic `identifier` field whose format depends on context — MSSV format

        for students, email format for staff.

        '
      properties:
        accountType:
          type: string
          enum:
          - STUDENT
          - STAFF
        studentId:
          type: string
          description: 'Required when `accountType=student`. MSSV format (e.g., `21127001`).

            '
          example: '21127001'
        email:
          type: string
          format: email
          description: Required when `accountType=staff`.
          example: btc01@unihub.edu.vn
        password:
          type: string
          format: password
          minLength: 8
      examples:
        student:
          value:
            accountType: STUDENT
            studentId: '21127001'
            password: '...'
        staff:
          value:
            accountType: STAFF
            email: btc01@unihub.edu.vn
            password: '...'
    LoginResponse:
      type: object
      required:
      - accessToken
      - tokenType
      - expiresIn
      - role
      - user
      description: '`refresh_token` is only populated in the JSON body for **mobile
        clients**.

        Web clients receive the refresh token via `Set-Cookie: HttpOnly` header instead.

        '
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum:
          - Bearer
          default: Bearer
        expiresIn:
          type: integer
          description: Access token lifetime in seconds (always 900)
          example: 900
        role:
          type: string
          enum:
          - STUDENT
          - BTC
          - CHECKIN_STAFF
        refreshToken:
          type: string
          nullable: true
          description: 'Present in body **only for mobile clients** (checkin_staff).
            `null` for web

            clients (web receives it via HttpOnly cookie). App stores this in

            Android Keystore / iOS Keychain.

            '
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      required:
      - id
      - role
      - fullName
      - email
      properties:
        id:
          type: string
          description: '`studentId` (TEXT) for students, UUID for staff. Frontend
            treats as opaque.

            '
        role:
          type: string
          enum:
          - STUDENT
          - BTC
          - CHECKIN_STAFF
        fullName:
          type: string
        email:
          type: string
          format: email
        allowedWorkshopIds:
          type: array
          items:
            type: string
            format: uuid
          description: Only populated for `checkin_staff`
    DeviceTokenRequest:
      type: object
      required:
      - token
      - platform
      properties:
        token:
          type: string
          minLength: 1
          maxLength: 4096
        platform:
          type: string
          enum:
          - IOS
          - ANDROID
    DeviceToken:
      type: object
      required:
      - id
      - platform
      - isActive
      - createdAt
      properties:
        id:
          type: string
          format: uuid
        platform:
          type: string
          enum:
          - IOS
          - ANDROID
        isActive:
          type: boolean
        lastSeen:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    AssignWorkshopsRequest:
      type: object
      required:
      - workshopIds
      properties:
        workshopIds:
          type: array
          items:
            type: string
            format: uuid
          description: Full replacement list of workshop UUIDs